Update Line Drawing Algorithm

+++ src/main/scala/Application/CoordinateMapper.scala
@@ -1,14 +1,14 @@
 package Application
 
 class CoordinateMapper {
-  private var canvasWidth  : Double = _
-  private var canvasHeight : Double = _
-  private var windowXMin : Double = _
-  private var windowXMax : Double = _
-  private var windowYMin : Double = _
-  private var windowYMax : Double = _
+  private var canvasWidth  : Int = _
+  private var canvasHeight : Int = _
+  private var windowXMin : Int = _
+  private var windowXMax : Int = _
+  private var windowYMin : Int = _
+  private var windowYMax : Int = _
 
-  def init(canvasWidth  : Double, canvasHeight : Double, windowXMin : Double, windowXMax : Double, windowYMin : Double, windowYMax : Double): Unit = {
+  def init(canvasWidth  : Int, canvasHeight : Int, windowXMin : Int, windowXMax : Int, windowYMin : Int, windowYMax : Int): Unit = {
     this.canvasWidth = canvasWidth
     this.canvasHeight = canvasHeight
     this.windowXMin = windowXMin
@@ -18,7 +18,7 @@ class CoordinateMapper {
   }
 
 
-  def mapToCanvasSpace(xWindow: Double, yWindow: Double): (Double, Double) = {
+  def mapToCanvasSpace(xWindow: Int, yWindow: Int): (Int, Int) = {
     val xCanvas = ((xWindow - windowXMin) / (windowXMax - windowXMin)) * canvasWidth
     val yCanvas = canvasHeight - ((yWindow - windowYMin) / (windowYMax - windowYMin)) * canvasHeight
 
@@ -30,7 +30,7 @@ class CoordinateMapper {
     val yCanvas = canvasHeight - ((yWindow - windowYMin) / (windowYMax - windowYMin)) * canvasHeight
     val radiusCanvas = (radius / (windowXMax - windowXMin)) * canvasWidth // You can also use canvasHeight if the aspect ratio is different
 
-    (xCanvas.toInt, yCanvas.toInt, radiusCanvas.toInt)
+    (xCanvas, yCanvas, radiusCanvas)
   }
